pyproject.toml
[project]
name = "pcdiff-implant"
version = "1.1.0"
description = "Point Cloud Diffusion Models for Automatic Implant Generation"
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Paul Friedrich"},
    {name = "Julia Wolleb"},
    {name = "Florentin Bieder"},
    {name = "Florian M. Thieringer"},
    {name = "Philippe C. Cattin"},
]

dependencies = [
    # Install PyTorch with CUDA 13 support separately, e.g.
    # uv pip install torch==2.5.0 --index-url https://download.pytorch.org/whl/cu124
    # uv pip install torchvision==0.20.0 --index-url https://download.pytorch.org/whl/cu124
    "torch>=2.4.0,<2.6.0",
    "torchvision>=0.19.0,<0.21.0",

    # Common dependencies for both models
    "matplotlib>=3.9.0",
    "tqdm>=4.66.0",
    "trimesh>=4.3.0",
    "open3d>=0.18.0",
    "pynrrd>=1.0.0",
    "PyMCubes>=0.1.4",
    "numpy>=2.1.0",
    "scipy>=1.14.0",

    # Voxelization-specific
    "pyyaml>=6.0",
    "tensorboard>=2.18.0",
    "plyfile>=1.0.3",
    "scikit-image>=0.23.0",
    "opencv-python>=4.10.0",
    "pykdtree>=1.3.7",
    "diplib>=3.4.0",
    "imageio>=2.33.0",
    "plotly>=5.23.0",

    # Point cloud diffusion specific
    "ninja>=1.11.1",
    "rich>=13.7.0",
    # Experiment tracking (required for training runs)
    "wandb>=0.16.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
wandb = ["wandb>=0.16.0"]

[dependency-groups]
dev = ["ipdb>=0.13.0", "pytest>=8.0.0", "pre-commit>=3.8.0", "ruff>=0.8.0"]

[tool.hatch.build.targets.wheel]
packages = ["pcdiff", "voxelization"]

[tool.hatch.build.targets.editable]
packages = ["pcdiff", "voxelization"]

[project.urls]
Homepage = "https://pfriedri.github.io/pcdiff-implant-io/"
Repository = "https://github.com/pfriedri/pcdiff-implant"
"Paper (arXiv)" = "https://arxiv.org/abs/2303.08061"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = [
    "E501",   # line too long — handled gradually
    "E402",   # module-import-not-at-top-of-file — common in research scripts with sys.path manipulation
    "E711",   # none-comparison — SQLAlchemy uses `== None`
    "E712",   # true-false-comparison — SQLAlchemy uses `== True/False`
    "E721",   # type-comparison — intentional in some cases
    "E722",   # bare-except — legacy research code
    "F401",   # unused imports — common in research code __init__.py
    "F403",   # star imports
    "F405",   # may be undefined from star import
    "F811",   # redefined-while-unused — common in notebook-style research code
    "F821",   # undefined-name — dynamic imports and conditional definitions
    "F841",   # unused-variable — common in research code (computed for debugging)
    "E741",   # ambiguous variable names (l, O, I) — common in math code
]

[tool.ruff.lint.per-file-ignores]
"pcdiff/modules/**" = ["E", "F"]  # CUDA extension bindings, leave as-is

[tool.pytest.ini_options]
testpaths = ["tests"]

